requirements.txt
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
jinja2==3.1.3
weasyprint==62.1
rich==13.7.1
PyYAML==6.0.1
